adlink-conf: Update local.conf to switch to build packagegroup-adlink and add PA_USER and PA_TOKEN for private repos

+++ conf/adlink-conf/adlink-imx-setup-release.sh
@@ -1,10 +1,25 @@
 #!/bin/sh
 
+TOKEN_PRE="ghp_"
+TOKEN_POST="n4138GgHZwaS7MChjf83NeQD2E7iUt01u9po"
+
 if [ -z "$BUILD" ]; then
 	BUILD=build
 fi
 
 MACHINE=$MACHINE DISTRO=$DISTRO source imx-setup-release.sh -b $BUILD
 
-cp ../sources/meta-adlink-nxp/conf/adlink-conf/$MACHINE/bblayers.conf ./conf/
-cp ../sources/meta-adlink-nxp/conf/adlink-conf/$MACHINE/local.conf ./conf/
+if [ -f ../sources/meta-adlink-nxp/conf/adlink-conf/$MACHINE/bblayers.conf ]; then
+	cp ../sources/meta-adlink-nxp/conf/adlink-conf/$MACHINE/bblayers.conf ./conf/
+else
+	cp ../sources/meta-adlink-nxp/conf/adlink-conf/$MACHINE/bblayers.conf.sample ./conf/bblayers.conf
+fi
+if [ -f ../sources/meta-adlink-nxp/conf/adlink-conf/$MACHINE/local.conf ]; then
+	cp ../sources/meta-adlink-nxp/conf/adlink-conf/$MACHINE/local.conf ./conf/
+else
+	cp ../sources/meta-adlink-nxp/conf/adlink-conf/$MACHINE/local.conf.sample ./conf/local.conf
+fi
+
+echo "PA_USER ?= \"adlink-guest\"" >> ./conf/local.conf
+echo "PA_TOKEN ?= \""${TOKEN_PRE}${TOKEN_POST}"\"" >> ./conf/local.conf
+
